requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-allauth==0.57.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
celery==5.3.4
redis==5.0.1
pillow==10.1.0
reportlab==4.4.3
weasyprint==60.2
pandas==2.1.4
numpy==1.25.2
python-decouple==3.8
django-extensions==3.2.3
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
